+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add genetically modified and edited experiment categories. [#509](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/509)
+- Add new status levels to proposals. [#504](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/504)
 
 ### Fixed