+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Move previous cropping to restrictions tab. [#512](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/512)
+- Clarify description for previous cropping. [#513](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/513)
 
 ### Fixed